+++ test/automated_test/data/fail_testmarkers_os3_no_init_obj.osexp
@@ -1,0 +1,107 @@
+---
+API: 2.1
+OpenSesame: 3.3.14
+Platform: nt
+---
+set width 1024
+set uniform_coordinates yes
+set title "New experiment"
+set subject_parity even
+set subject_nr 0
+set start experiment
+set sound_sample_size -16
+set sound_freq 48000
+set sound_channels 2
+set sound_buf_size 1024
+set round_decimals 2
+set height 768
+set fullscreen no
+set form_clicks no
+set foreground white
+set font_underline no
+set font_size 18
+set font_italic no
+set font_family mono
+set font_bold no
+set experiment_path "D:/opensesame3_plugin_markers/test/data"
+set disable_garbage_collection yes
+set description "The main experiment item"
+set coordinates uniform
+set compensation 0
+set canvas_backend psycho
+set background black
+
+define sequence experiment
+	set flush_keyboard yes
+	set description "Runs a number of items in sequence"
+	run welcome always
+	run new_loop always
+
+define sketchpad fixation
+	set duration 100
+	set description "Displays stimuli"
+	draw fixdot color=white show_if=always style=default x=0 y=0 z_index=0
+
+define logger new_logger
+	set description "Logs experimental data"
+	set auto_log yes
+
+define loop new_loop
+	set source_file ""
+	set source table
+	set repeat 1
+	set order sequential
+	set description "Repeatedly runs another item"
+	set cycles 10
+	set continuous no
+	set break_if_on_first yes
+	set break_if never
+	setcycle 0 marker_value 1
+	setcycle 0 marker_duration 10
+	setcycle 1 marker_value 2
+	setcycle 1 marker_duration 10
+	setcycle 2 marker_value 3
+	setcycle 2 marker_duration 10
+	setcycle 3 marker_value 4
+	setcycle 3 marker_duration 10
+	setcycle 4 marker_value 5
+	setcycle 4 marker_duration 10
+	setcycle 5 marker_value 6
+	setcycle 5 marker_duration 10
+	setcycle 6 marker_value 7
+	setcycle 6 marker_duration 10
+	setcycle 7 marker_value 8
+	setcycle 7 marker_duration 10
+	setcycle 8 marker_value 9
+	setcycle 8 marker_duration 10
+	setcycle 9 marker_value 10
+	setcycle 9 marker_duration 10
+	run new_sequence
+
+define markers_os3_send new_markers_os3_send
+	set marker_value "[marker_value]"
+	set marker_reset_to_zero yes
+	set marker_object_duration "[marker_duration]"
+	set marker_device_tag marker_device_1
+	set description "Sends marker to Leiden Univ marker device - Markers plugin for OpenSesame 3"
+
+define sequence new_sequence
+	set flush_keyboard yes
+	set description "Runs a number of items in sequence"
+	run fixation always
+	run stimulus always
+	run new_markers_os3_send always
+	run new_logger always
+
+define sketchpad stimulus
+	set duration 0
+	set description "Displays stimuli"
+	draw textline center=1 color=white font_bold=no font_family=mono font_italic=no font_size=18 html=yes show_if=always text="Sending marker [marker_value]<br /><br />Press any key to continue. Press esq to exit." x=0 y=0 z_index=0
+
+define sketchpad welcome
+	set start_response_interval no
+	set reset_variables no
+	set duration 100
+	set description "Displays stimuli"
+	draw textline center=1 color=white font_bold=no font_family=serif font_italic=no font_size=32 html=yes show_if=always text="OpenSesame 3.3 <i>Lentiform Loewenfeld</i>" x=0 y=0 z_index=0
+
